+++ examples/simple-usage/index.ts
@@ -1,5 +1,6 @@
 import "reflect-metadata";
 import { ApolloServer } from "apollo-server";
+import * as path from "path";
 import { buildSchema } from "../../src";
 
 import { RecipeResolver } from "./recipe-resolver";
@@ -8,6 +9,8 @@ async function bootstrap() {
   // build TypeGraphQL executable schema
   const schema = await buildSchema({
     resolvers: [RecipeResolver],
+    // automatically create `schema.gql` file with schema definition in current folder
+    emitSchemaFile: path.resolve(__dirname, "schema.gql"),
   });
 
   // Create GraphQL server